Use swarm for color mixing

+++ src/main/java/es/sixey/png2svg/color/Mix.java
@@ -1,3 +1,3 @@
 package es.sixey.png2svg.color;
 
-public record Mix(Color color, Float weight){}
+public record Mix(Color color, Double weight){}
